requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
redis[asyncio]==5.0.0
apscheduler==3.10.4
pydantic==2.7.0
pydantic-settings==2.3.0
httpx==0.27.0
numpy==1.26.0
pandas>=2.0.0
yfinance>=0.2.38
scikit-learn==1.5.0
sentence-transformers==3.0.0
python-dotenv==1.0.0
pytest==8.2.0
pytest-asyncio==0.23.0